+++ src/MCUScope/Dialogs/CommunicationSettingsDialog.xaml.cs
@@ -1,27 +1,48 @@
 using MCUScope.ViewModels;
 using System.Globalization;
 using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Controls.Primitives;
 
 namespace MCUScope.Dialogs
 {
     public partial class CommunicationSettingsDialog : Window
     {
         private readonly MainViewModel _vm;
+        private static readonly int[] CommonBaudRates =
+        {
+            9600, 19200, 38400, 57600, 115200, 230400, 460800, 921600,
+            1000000, 1500000, 2000000, 3000000
+        };
 
         public CommunicationSettingsDialog(MainViewModel vm)
         {
             InitializeComponent();
             _vm = vm;
-            BaudRateTextBox.Text = vm.Settings.Communication.BaudRate.ToString(CultureInfo.InvariantCulture);
+
+            foreach (int baud in CommonBaudRates)
+            {
+                BaudRateComboBox.Items.Add(baud.ToString(CultureInfo.InvariantCulture));
+            }
+
+            BaudRateComboBox.Text = vm.Settings.Communication.BaudRate.ToString(CultureInfo.InvariantCulture);
             ClockTextBox.Text = vm.Settings.Communication.BaseClockMHz.ToString("F2", CultureInfo.InvariantCulture);
             UpdateRateText();
             ClockTextBox.TextChanged += (s, e) => UpdateRateText();
-            BaudRateTextBox.TextChanged += (s, e) => UpdateRateText();
+            BaudRateComboBox.AddHandler(TextBoxBase.TextChangedEvent,
+                new TextChangedEventHandler(OnBaudRateTextChanged));
+            BaudRateComboBox.SelectionChanged += (s, e) => UpdateRateText();
+            BaudRateComboBox.LostKeyboardFocus += (s, e) => UpdateRateText();
+        }
+
+        private void OnBaudRateTextChanged(object sender, TextChangedEventArgs e)
+        {
+            UpdateRateText();
         }
 
         private void UpdateRateText()
         {
-            if (int.TryParse(BaudRateTextBox.Text, NumberStyles.Integer, CultureInfo.InvariantCulture, out var baudRate) &&
+            if (int.TryParse(BaudRateComboBox.Text, NumberStyles.Integer, CultureInfo.InvariantCulture, out var baudRate) &&
                 baudRate > 0)
             {
                 RateText.Text = $"Using baud rate: {baudRate} bps ({baudRate / 1_000_000.0:F3} Mbps)";
@@ -41,7 +62,7 @@ namespace MCUScope.Dialogs
 
         private void OnOk(object sender, RoutedEventArgs e)
         {
-            bool baudOk = int.TryParse(BaudRateTextBox.Text, NumberStyles.Integer, CultureInfo.InvariantCulture, out var baudRate) &&
+            bool baudOk = int.TryParse(BaudRateComboBox.Text, NumberStyles.Integer, CultureInfo.InvariantCulture, out var baudRate) &&
                           baudRate > 0;
             bool clockOk = double.TryParse(ClockTextBox.Text, NumberStyles.Float, CultureInfo.InvariantCulture, out var clock) &&
                            clock > 0;
@@ -55,6 +76,7 @@ namespace MCUScope.Dialogs
 
             _vm.Settings.Communication.BaudRate = baudOk ? baudRate : 0;
             _vm.Settings.Communication.BaseClockMHz = clockOk ? clock : 0;
+            _vm.NotifyCommunicationSettingsChanged();
             DialogResult = true;
             Close();
         }